+++ src/main/java/ru/hogwarts/school/service/StudentService.java
@@ -6,9 +6,9 @@ import java.util.Collection;
 
 public interface StudentService {
     public  Student createStudent(Student student);
-    public Student findStudent(int id);
+    public Student findStudent(long id);
     public Student updateStudent(Student student);
-    public void deleteStudent(int id);
+    public void deleteStudent(long id);
     public Collection<Student> getAll();
     public Collection<Student> findByAge(int age);
     public Collection<Student> findByAgeBetween(int minAge, int maxAge);
